requirements.txt
asgiref==3.3.1
certifi==2020.12.5
chardet==3.0.4
Django==3.1.4
django-filter==2.4.0
djangorestframework==3.12.2
idna==2.10
importlib-metadata==3.1.1
Markdown==3.3.3
pkg-resources==0.0.0
psycopg2==2.8.6
python-decouple==3.3
pytz==2020.4
requests==2.25.0
sqlparse==0.4.1
urllib3==1.26.2
zipp==3.4.0
